requirements.txt
--extra-index-url https://google-coral.github.io/py-repo/
aiofiles==0.7.0
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
fastapi==0.66.0
h11==0.12.0
httptools==0.2.0
numpy==1.21.0
Pillow==7.2.0
pydantic==1.8.2
python-dotenv==0.18.0
PyYAML==5.4.1
starlette==0.14.2
tflite-runtime==2.5.0
typing-extensions==3.10.0.0
uvicorn==0.14.0
watchgod==0.7
websockets==9.1
